requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.3
kaggle==1.5.15